+++ shared/r2-shared-swift/Toolkit/Extensions/URL.swift
@@ -13,6 +13,11 @@ import Foundation
 import CommonCrypto
 
 extension URL: Loggable {
+
+    /// Indicates whether this URL is an HTTP or HTTPS URL.
+    public var isHTTP: Bool {
+        ["http", "https"].contains(scheme?.lowercased())
+    }
     
     /// Returns whether the given `url` is `self` or one of its descendants.
     public func isParentOf(_ url: URL) -> Bool {